Revert "serwer, poczatek GUI, refaktoryzacja"
