--- GameRequest.java
@@ -1,4 +1,0 @@
-public enum GameRequest
-{
-    JOIN, LEAVE, KICK, OK, WORLD, MOVE_UP, MOVE_DOWN, MOVE_LEFT, MOVE_RIGHT;
-}
